[REFACTOR] #258 Posts/Comments 도메인 분리 및 공유 컴포넌트 이동
