+++ src/domains/Posts/hooks/useBulkDeletePost.ts
