+++ src/domains/Posts/hooks/useDeletePost.ts
