+++ src/domains/Posts/types/post.ts
