+++ src/shared/components/BulkDeleteBar.tsx
