+++ src/shared/components/MemberInfoPopover.tsx
