+++ src/shared/components/index.ts
@@ -1,6 +1,8 @@
 export { AppSidebar } from './AppSidebar';
+export { default as BulkDeleteBar } from './BulkDeleteBar';
 export { DateTimePicker } from './DateTimePicker';
 export { default as Header } from './Header';
+export { default as MemberInfoPopover } from './MemberInfoPopover';
 export { NavUser } from './NavUser';
 export { default as PageHeader } from './PageHeader';
 export { PeriodStatusBadge } from './PeriodStatusBadge';
